Splut drawable position from drawable node.
